+++ packages/core/loglayer/src/MockLogLayer.ts
@@ -68,11 +68,9 @@ export class MockLogLayer implements ILogLayer<MockLogLayer>, ILogBuilder<MockLo
     return this.mockLogBuilder.withMetadata(_metadata) as any;
   }
 
-  getContext(_options?: { evalLazy?: boolean }): Record<string, any> {
-    return {};
-  }
-
-  async getContextAsync(_options?: { evalLazy?: boolean }): Promise<Record<string, any>> {
+  getContext(): Record<string, any>;
+  getContext(_options: { evalLazy?: boolean }): Record<string, any> | Promise<Record<string, any>>;
+  getContext(_options?: { evalLazy?: boolean }): Record<string, any> | Promise<Record<string, any>> {
     return {};
   }